Added incident update mechanism, added request_delay var for Apps
