+++ app/im/mattermost/config.py
@@ -29,6 +29,7 @@ buttons = {
         }
     }
 }
+mattermost_request_delay = 0.5  # seconds
 
 
 def mattermost_bold_text(value):
