+++ app/im/mattermost/messages.py
@@ -3,7 +3,7 @@ from time import sleep
 
 import requests
 
-from .config import mattermost_headers
+from .config import mattermost_headers, mattermost_request_delay
 
 
 def mattermost_send_message(url, channel_id, message):
@@ -15,5 +15,5 @@ def mattermost_send_message(url, channel_id, message):
         'unfurl_media': False
     }
     response = requests.post(f'{url}/api/chat.postMessage', headers=mattermost_headers, data=json.dumps(payload))
-    sleep(0.1)
+    sleep(mattermost_request_delay)
     return response.json().get('ts')
