+++ app/im/message_template.py
@@ -1,8 +1,8 @@
 from jinja2 import Template
 
-from app.logging import logger
 from app.im.mattermost.message_template import default_message_template as mattermost_default_message_template
 from app.im.slack.message_template import default_message_template as slack_default_message_template
+from app.logging import logger
 
 
 class MessageTemplate:
