+++ app/im/slack/__init__.py
@@ -1,4 +1,4 @@
 from .buttons import reformat_message
-from .config import buttons
 from .channels import slack_get_public_channels
+from .config import buttons
 from .messages import slack_send_message
