+++ app/im/slack/messages.py
@@ -1,8 +1,9 @@
 import json
+from time import sleep
 
 import requests
 
-from .config import slack_headers
+from .config import slack_headers, slack_request_delay
 
 
 def slack_send_message(url, channel_id, message):
@@ -14,4 +15,5 @@ def slack_send_message(url, channel_id, message):
         'unfurl_media': False
     }
     response = requests.post(f'{url}/api/chat.postMessage', headers=slack_headers, data=json.dumps(payload))
+    sleep(slack_request_delay)
     return response.json().get('ts')
