+++ app/logging.py
@@ -1,4 +1,5 @@
 import logging
+
 from config import log_level
