+++ app/time.py
@@ -1,0 +1,8 @@
+from datetime import timedelta
+
+
+def unix_sleep_to_timedelta(unix_sleep_time):
+    value = int(unix_sleep_time[:-1])
+    unit = unix_sleep_time[-1]
+    unit_map = {'s': 'seconds', 'm': 'minutes', 'h': 'hours', 'd': 'days'}
+    return timedelta(**{unit_map[unit]: value})
